refactor: moved RegexListViewer to ui package
